removed showChannel from actions
