+++ react/constants/show_action_types.js
@@ -11,7 +11,7 @@ export const ASSET_NEW_ASYNC = 'ASSET_NEW_ASYNC';
 export const ASSET_NEW_SUCCESS = `ASSET_NEW_SUCCESS`;
 
 // channel actions
-export const CHANNEL_REQUEST_ASYNC = 'CHANNEL_REQUEST_ASYNC';
+export const CHANNEL_REQUEST_NEW = 'CHANNEL_REQUEST_NEW';
 export const CHANNEL_REQUEST_SUCCESS = 'CHANNEL_REQUEST_SUCCESS';
 
 export const CHANNEL_NEW_ASYNC = 'CHANNEL_NEW_ASYNC';
