+++ react/sagas/index.js
@@ -1,14 +1,13 @@
 import { all } from 'redux-saga/effects';
-import { watchNewAssetRequest, watchNewChannelRequest } from './request';
+import { watchNewAssetRequest } from './show_asset';
+import { watchNewChannelRequest, watchUpdateChannelClaims } from './show_channel';
 import { watchFileIsRequested } from './file';
-import { watchShowNewChannel, watchShowNewChannelClaimsRequest } from './show_channel';
 
 export default function* rootSaga () {
   yield all([
     watchNewAssetRequest(),
     watchNewChannelRequest(),
-    watchShowNewChannel(),
+    watchUpdateChannelClaims(),
     watchFileIsRequested(),
-    watchShowNewChannelClaimsRequest(),
   ]);
 }
